Install script Github fixes
In v3 the config dir of the frontend no longer exists.
The config option 'nfsen.config-file' is also absent.
The regex "'backend.path'] =" matches nothing, "'backend.path'] " does.

Originally the script would download the tarball if the tarball would be
absent but the contents were present. Now it also checks for the
contents. I.e. a simple clone of the repo suffices.